Shaders added

+++ apps/forward-renderer/Application.hpp
@@ -13,8 +13,8 @@ public:
 
     int run();
 private:
-    const size_t m_nWindowWidth = 1280;
-    const size_t m_nWindowHeight = 720;
+    const int m_nWindowWidth = 1280;
+    const int m_nWindowHeight = 720;
     glmlv::GLFWHandle m_GLFWHandle{ m_nWindowWidth, m_nWindowHeight, "Template" }; // Note: the handle must be declared before the creation of any object managing OpenGL resource (e.g. GLProgram, GLShader)
 
     const glmlv::fs::path m_AppPath;
